+++ ci/install_osx.sh
@@ -1,20 +1,6 @@
-brew tap dartsim/dart # for ipopt
+#!/usr/bin/env bash
+ 
+set -e
 
 brew update > /dev/null
-
-brew install git
-brew install cmake
-brew install assimp
-brew install fcl
-brew install bullet
-brew install ode
-brew install flann
-brew install boost
-brew install eigen
-brew install tinyxml
-brew install tinyxml2
-brew install libccd
-brew install nlopt
-brew install ipopt
-brew install urdfdom
-brew install open-scene-graph
+brew bundle