+++ dart/common/Memory.hpp
@@ -39,9 +39,7 @@
 
 #include "dart/config.hpp"
 #include "dart/common/Deprecated.hpp"
-
-#include "dart/config.hpp"
-#include "dart/common/Deprecated.hpp"
+#include "dart/common/detail/AlignedAllocator.hpp"
 
 namespace dart {
 namespace common {